+++ src/features/language-choice/language-data.js
@@ -1,0 +1,281 @@
+export const LANGUAGE_SETTINGS = {
+    default: null,
+
+    dragonlance: {
+        standard: {
+            add: {
+                abanasinian: 'Abanasinian',
+                camptalk: 'Camptalk',
+                ergot: 'Ergot',
+                kenderspeak: 'Kenderspeak',
+                kharolian: 'Kharolian',
+                khur: 'Khur',
+                nordmaarian: 'Nordmaarian',
+                solamnic: 'Solamnic',
+                elven: 'Elven',
+                dwarven: 'Dwarven',
+                plainsfolk: 'Plainsfolk'
+            },
+            remove: ['draconic', 'giant', 'goblin', 'halfling', 'orc', 'sign', 'elvish', 'dwarvish']
+        },
+        exotic: {
+            add: {
+                draconic: 'Draconic',
+                dargoi: 'Dargoi',
+                goblin: 'Goblin',
+                gullytalk: 'Gullytalk',
+                istarian: 'Istarian',
+                kothian: 'Kothian',
+                kerakese: 'Kerakese',
+                magius: 'Magius',
+                ogre: 'Ogre',
+                primordial: 'Primordial',
+                sylvan: 'Sylvan',
+                panak: 'Panak',
+                thanoi: 'Thanoi'
+            },
+            remove: ['aarakocra', 'abyssal', 'celestial', 'deep', 'gith', 'gnoll', 'infernal', 'undercommon']
+        }
+    },
+
+    severedlands: {
+        standard: {
+            add: {
+                zlonyjun: 'Zlonyjun (Common)',
+                tilgi: 'Tilgi',
+                ocheen: 'Ocheen',
+                zimhikhi: 'Zimhikhi',
+                zhanyboian: 'Zhanyboian',
+                nadfian: 'Nadfīan',
+                frui: 'Frui'
+            },
+            remove: ['common', 'draconic', 'dwarvish', 'elvish', 'gnomish', 'halfling', 'aarakocra', 'gith', 'gnoll', 'goblin', 'orc']
+        },
+        exotic: {
+            add: {},
+            remove: []
+        }
+    },
+
+    forgottenrealms: {
+        standard: {
+            add: {
+                // Major Racial Languages
+                dwarvish: 'Dwarvish/Dethek',
+                davek: 'Davek',
+                elven: 'Elven',
+                espruar: 'Espruar',
+                seaelvish: 'Sea Elvish',
+                gnomish: 'Gnomish/Gnim',
+                halfling: 'Halfling/Luiric',
+
+                // Thorass Languages
+                common: 'Common',
+                chondathan: 'Chondathan',
+                maidenstongue: 'Maiden\'s Tongue',
+                tharian: 'Tharian',
+                telpi: 'Telpi',
+                turmic: 'Turmic',
+                aglarondan: 'Aglarondan',
+
+                // Chessan Languages
+                chessentan: 'Chessentan',
+                sespechian: 'Sespechian',
+                shaaran: 'Shaaran',
+                tashalan: 'Tashalan',
+
+                // Illuski Languages
+                illuskan: 'Illuskan',
+                bothii: 'Bothii',
+                dambrathan: 'Dambrathan',
+                reghedjic: 'Reghedjic',
+
+                // Ulou Languages
+                damaran: 'Damaran',
+                easting: 'Easting',
+                andt: 'Andt',
+                dtarig: 'D\'tarig',
+                erakic: 'Erakic',
+                halruaan: 'Halruaan',
+                uluik: 'Uluik',
+                ulutiun: 'Ulutiun',
+                naric: 'Naric',
+                ruathlek: 'Ruathlek',
+
+                // Raumtheran Languages
+                rashemi: 'Rashemi',
+                allesian: 'Allesian',
+                sossic: 'Sossic',
+
+                // Rauric Languages
+                untheric: 'Untheric',
+                alzhedo: 'Alzhedo',
+                kadari: 'Kadari',
+                maran: 'Maran',
+                midani: 'Midani',
+                mulhorandi: 'Mulhorandi',
+                muhjuri: 'Muhjuri',
+                thayan: 'Thayan',
+
+                // Imaskari Languages
+                chuchian: 'Chuchian',
+                commani: 'Commani',
+                dalat: 'Dalat',
+                durpari: 'Durpari',
+                durparishaartan: 'Durpari-Shaartan Patois',
+                gur: 'Gur',
+                gurri: 'Gurri',
+                khassidi: 'Khassidi',
+                lantanese: 'Lantanese',
+                naican: 'Naican',
+                oigur: 'Oigur',
+                quirish: 'Quirish',
+                raumvira: 'Raumvira',
+                raurindi: 'Raurindi',
+                tuigan: 'Tuigan',
+                zamogedi: 'Zamogedi',
+
+                // Kara-Turan Languages
+                bavanese: 'Bavanese',
+                bawani: 'Bawani',
+                bertanese: 'Bertanese',
+                han: 'Han',
+                highshou: 'High Shou',
+                issacortae: 'Issacortae',
+                khazari: 'Khazari',
+                koryo: 'Koryo',
+                kozakuran: 'Kozakuran',
+                kuong: 'Kuong',
+                laothan: 'Laothan',
+                lidahan: 'Lidahan',
+                pazruki: 'Pazruki',
+                purang: 'Purang',
+                rakhati: 'Ra-Khati',
+                tulung: 'T\'u Lung',
+                tabotan: 'Tabotan',
+                tradetongue: 'Trade Tongue',
+                waan: 'Wa-an',
+                wuhaltai: 'Wu-haltai',
+
+                // Other Regional
+                chultan: 'Chultan',
+                waelan: 'Waelan',
+                cosh: 'Cosh'
+            },
+            remove: ['common', 'dwarvish', 'elvish', 'giant', 'gnomish', 'goblin', 'halfling', 'orc', 'draconic', 'sign']
+        },
+        exotic: {
+            add: {
+                // Draconic Languages
+                aragrakh: 'Aragrakh',
+                auldwyrmish: 'Auld Wyrmish',
+                draconic: 'Draconic',
+                draconicsign: 'Draconic Sign Language',
+                iokharic: 'Iokharic',
+                olddraconic: 'Old Draconic',
+                troglodyte: 'Troglodyte',
+                tymantheran: 'Tymantheran',
+                yipyak: 'Yipyak',
+
+                // Exotic Elven
+                drow: 'Drow',
+                drowsign: 'Drow Sign Language',
+                hamarfae: 'Hamarfae',
+                highwealdan: 'High Wealdan',
+                rellanic: 'Rellanic',
+                seldruin: 'Seldruin',
+                tuilvilanuue: 'Tuilvilanuue',
+
+                // Orcish/Goblin/Giant
+                orcish: 'Orcish/Daraktan',
+                goblin: 'Goblin/Ghukliak',
+                giant: 'Giant/Jotun',
+                ogre: 'Ogre',
+
+                // Gith
+                gith: 'Gith',
+
+                // Ancient Thorass
+                thorass: 'Thorass',
+                jhaamdathan: 'Jhaamdathan',
+                auldtharian: 'Auld Tharian',
+
+                // Ancient Chessan
+                auldchessic: 'Auld Chessic',
+                akalan: 'Akalan',
+                alambit: 'Alambit',
+                eshowan: 'Eshowan',
+                telfir: 'Telfir',
+
+                // Ancient Illuski
+                oldilluskan: 'Old Illuskan',
+
+                // Ancient Ulou
+                loross: 'Loross',
+                narfelli: 'Narfelli',
+                netherese: 'Netherese',
+                rengardt: 'Rengardt',
+
+                // Ancient Raumtheran
+                halardrim: 'Halardrim',
+
+                // Ancient Rauric
+                noga: 'Noga',
+
+                // Ancient Imaskari
+                devic: 'Devic',
+                roushoum: 'Roushoum',
+                ancientimask: 'Ancient Imask',
+
+                // Other Ancient/Dead
+                ancientsalamic: 'Ancient Salamic',
+                affa: 'Affa',
+                azuposi: 'Azuposi',
+                chun: 'Chun',
+                oldanwahran: 'Old Anwahran',
+                oldomuan: 'Old Omuan',
+                nexalan: 'Nexalan',
+                payit: 'Payit',
+                susrit: 'Susrit',
+
+                // Planar/Elemental
+                abyssal: 'Abyssal',
+                infernal: 'Infernal',
+                celestial: 'Celestial',
+                aquan: 'Aquan',
+                auran: 'Auran',
+                ignan: 'Ignan',
+                terran: 'Terran',
+                primordial: 'Primordial',
+
+                // Fey Languages
+                sylvan: 'Sylvan',
+                brownie: 'Brownie',
+                dryad: 'Dryad',
+                nereid: 'Nereid',
+                nixie: 'Nixie',
+                nymph: 'Nymph',
+                pixie: 'Pixie',
+                satyr: 'Satyr',
+                sirine: 'Sirine',
+                sprite: 'Sprite',
+                sylph: 'Sylph',
+
+                // Aberration
+                qualith: 'Qualith',
+                beholder: 'Beholder/Uibilaqthraxx',
+
+                // Underdark
+                undercommon: 'Undercommon',
+
+                // Other
+                drueidan: 'Drueidan',
+                daelic: 'Daelic',
+                deepSpeech: 'Deep Speech',
+                aarakocra: 'Aarakocra',
+                gnoll: 'Gnoll'
+            },
+            remove: ['aarakocra', 'abyssal', 'celestial', 'deep', 'draconic', 'gith', 'gnoll', 'infernal', 'primordial', 'sylvan', 'undercommon']
+        }
+    }
+};